requirements.txt
langchain==0.1.13
langchain-community
langchain_huggingface
python-dotenv
streamlit==1.32.2
streamlit-chat==0.1.1
tiktoken==0.5.2
faiss-cpu==1.8.0
sentence_transformers
